LPD-75170 refactor: rename ai-hub-rest-impl and ai-hub-rest-api to ai-hub-nexus-rest-impl and ai-hub-nexus-rest-api
