+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-impl/src/main/java/com/liferay/ai/hub/nexus/internal/agent/SupervisorAgentImpl.java
@@ -8,7 +8,7 @@ package com.liferay.ai.hub.nexus.internal.agent;
 import com.liferay.ai.hub.nexus.agent.AgentContext;
 import com.liferay.ai.hub.nexus.agent.SupervisorAgent;
 import com.liferay.ai.hub.nexus.internal.memory.ChatMemoryProviderUtil;
-import com.liferay.ai.hub.rest.resource.v1_0.util.SseUtil;
+import com.liferay.ai.hub.nexus.rest.resource.v1_0.util.SseUtil;
 import com.liferay.object.rest.dto.v1_0.ObjectEntry;
 import com.liferay.object.rest.manager.v1_0.ObjectEntryManager;
 import com.liferay.object.service.ObjectDefinitionLocalService;
