--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-api/src/main/resources/com/liferay/ai/hub/rest/resource/v1_0/util/packageinfo
@@ -1,1 +1,0 @@
-version 3.0.0
