+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/resource/v1_0/AgentDefinitionResourceImpl.java
@@ -5,9 +5,9 @@
 
 package com.liferay.ai.hub.rest.internal.resource.v1_0;
 
-import com.liferay.ai.hub.rest.dto.v1_0.AgentDefinition;
-import com.liferay.ai.hub.rest.manager.v1_0.AgentDefinitionManager;
-import com.liferay.ai.hub.rest.resource.v1_0.AgentDefinitionResource;
+import com.liferay.ai.hub.nexus.rest.dto.v1_0.AgentDefinition;
+import com.liferay.ai.hub.nexus.rest.manager.v1_0.AgentDefinitionManager;
+import com.liferay.ai.hub.nexus.rest.resource.v1_0.AgentDefinitionResource;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.search.filter.Filter;
