+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/resource/v1_0/BaseMessageResourceImpl.java
@@ -5,8 +5,8 @@
 
 package com.liferay.ai.hub.rest.internal.resource.v1_0;
 
-import com.liferay.ai.hub.rest.dto.v1_0.Message;
-import com.liferay.ai.hub.rest.resource.v1_0.MessageResource;
+import com.liferay.ai.hub.nexus.rest.dto.v1_0.Message;
+import com.liferay.ai.hub.nexus.rest.resource.v1_0.MessageResource;
 import com.liferay.petra.function.UnsafeFunction;
 import com.liferay.petra.function.transform.TransformUtil;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
