+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/resource/v1_0/TaskResourceImpl.java
@@ -6,11 +6,11 @@
 package com.liferay.ai.hub.rest.internal.resource.v1_0;
 
 import com.liferay.account.model.AccountEntry;
+import com.liferay.ai.hub.nexus.rest.dto.v1_0.Task;
+import com.liferay.ai.hub.nexus.rest.resource.v1_0.TaskResource;
+import com.liferay.ai.hub.nexus.rest.resource.v1_0.util.SseUtil;
 import com.liferay.ai.hub.nexus.util.AccountEntryUtil;
-import com.liferay.ai.hub.rest.dto.v1_0.Task;
 import com.liferay.ai.hub.rest.internal.resource.v1_0.util.WorkflowContextUtil;
-import com.liferay.ai.hub.rest.resource.v1_0.TaskResource;
-import com.liferay.ai.hub.rest.resource.v1_0.util.SseUtil;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.workflow.WorkflowDefinition;
 import com.liferay.portal.kernel.workflow.WorkflowInstance;
