+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/agent-definition.properties
@@ -3,10 +3,10 @@
 #
 
 api.version=v1.0
-batch.engine.entity.class.name=com.liferay.ai.hub.rest.dto.v1_0.AgentDefinition
+batch.engine.entity.class.name=com.liferay.ai.hub.nexus.rest.dto.v1_0.AgentDefinition
 batch.engine.task.item.delegate=true
 batch.planner.export.enabled=true
 batch.planner.import.enabled=false
-entity.class.name=com.liferay.ai.hub.rest.dto.v1_0.AgentDefinition
+entity.class.name=com.liferay.ai.hub.nexus.rest.dto.v1_0.AgentDefinition
 osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.AI.Hub.REST)
 osgi.jaxrs.resource=true
