+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/chat.properties
@@ -3,6 +3,6 @@
 #
 
 api.version=v1.0
-entity.class.name=com.liferay.ai.hub.rest.dto.v1_0.Chat
+entity.class.name=com.liferay.ai.hub.nexus.rest.dto.v1_0.Chat
 osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.AI.Hub.REST)
 osgi.jaxrs.resource=true
