+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/task.properties
@@ -3,10 +3,10 @@
 #
 
 api.version=v1.0
-batch.engine.entity.class.name=com.liferay.ai.hub.rest.dto.v1_0.Task
+batch.engine.entity.class.name=com.liferay.ai.hub.nexus.rest.dto.v1_0.Task
 batch.engine.task.item.delegate=true
 batch.planner.export.enabled=false
 batch.planner.import.enabled=true
-entity.class.name=com.liferay.ai.hub.rest.dto.v1_0.Task
+entity.class.name=com.liferay.ai.hub.nexus.rest.dto.v1_0.Task
 osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.AI.Hub.REST)
 osgi.jaxrs.resource=true
