+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-test/src/testIntegration/java/com/liferay/ai/hub/rest/resource/v1_0/test/BaseAgentDefinitionResourceTestCase.java
@@ -912,7 +912,7 @@ public abstract class BaseAgentDefinitionResourceTestCase {
 
 		for (java.lang.reflect.Field field :
 				getDeclaredFields(
-					com.liferay.ai.hub.rest.dto.v1_0.AgentDefinition.class)) {
+					com.liferay.ai.hub.nexus.rest.dto.v1_0.AgentDefinition.class)) {
 
 			if (!ArrayUtil.contains(
 					getAdditionalAssertFieldNames(), field.getName())) {
@@ -1723,7 +1723,7 @@ public abstract class BaseAgentDefinitionResourceTestCase {
 	private com.liferay.portal.kernel.model.User _testCompanyAdminUser;
 
 	@Inject
-	private com.liferay.ai.hub.rest.resource.v1_0.AgentDefinitionResource
+	private com.liferay.ai.hub.nexus.rest.resource.v1_0.AgentDefinitionResource
 		_agentDefinitionResource;
 
 }
